Backend (FastAPI)/requirements.txt
pymongo
fastapi
passlib
python-jose
python-multipart